+++ matrix-sdk-android/src/main/java/im/vector/matrix/android/api/session/homeserver/HomeServerCapabilitiesService.kt
@@ -14,17 +14,16 @@
  * limitations under the License.
  */
 
-package im.vector.matrix.android.internal.network
+package im.vector.matrix.android.api.session.homeserver
 
-internal object NetworkConstants {
+/**
+ * This interface defines a method to retrieve the homeserver capabilities.
+ */
+interface HomeServerCapabilitiesService {
 
-    private const val URI_API_PREFIX_PATH = "_matrix/client"
-    const val URI_API_PREFIX_PATH_R0 = "$URI_API_PREFIX_PATH/r0/"
-    const val URI_API_PREFIX_PATH_UNSTABLE = "$URI_API_PREFIX_PATH/unstable/"
-
-
-    // Media
-    private const val URI_API_MEDIA_PREFIX_PATH = "_matrix/media"
-    const val URI_API_MEDIA_PREFIX_PATH_R0 = "$URI_API_MEDIA_PREFIX_PATH/r0/"
+    /**
+     * Get the HomeServer capabilities
+     */
+    fun getHomeServerCapabilities(): HomeServerCapabilities
 
 }