+++ vector/src/main/java/im/vector/riotx/core/utils/TextUtils.kt
@@ -16,6 +16,9 @@
 
 package im.vector.riotx.core.utils
 
+import android.content.Context
+import android.os.Build
+import android.text.format.Formatter
 import java.util.*
 
 object TextUtils {
@@ -40,6 +43,30 @@ object TextUtils {
             return if (hasDecimal) "${truncated / 10.0}$suffix" else "${truncated / 10}$suffix"
         } catch (t: Throwable) {
             return value.toString()
+        }
+    }
+
+    /**
+     * Since Android O, the system considers that 1ko = 1000 bytes instead of 1024 bytes. We want to avoid that for the moment.
+     */
+    fun formatFileSize(context: Context, sizeBytes: Long, useShortFormat: Boolean = false): String {
+        val normalizedSize = if (Build.VERSION.SDK_INT <= Build.VERSION_CODES.N) {
+            sizeBytes
+        } else {
+            // First convert the size
+            when {
+                sizeBytes < 1024               -> sizeBytes
+                sizeBytes < 1024 * 1024        -> sizeBytes * 1000 / 1024
+                sizeBytes < 1024 * 1024 * 1024 -> sizeBytes * 1000 / 1024 * 1000 / 1024
+                else                           -> sizeBytes * 1000 / 1024 * 1000 / 1024 * 1000 / 1024
+            }
+        }
+
+        return if (useShortFormat) {
+            Formatter.formatShortFileSize(context, normalizedSize)
+        } else {
+            Formatter.formatFileSize(context, normalizedSize)
+
         }
     }
 }